Refactor: extract shared OpenAI-compatible backend logic
- Add openai_compat.rs with OpenAiSseParser (renamed from ZaiSseParser),
  shared build_request_body() returning Result to propagate serialization
  errors instead of silently injecting null, and send_request() helper
- Reduce ZaiBackend and OllamaBackend to thin wrappers (~30 lines each)
- Update tests/zai_test.rs to use OpenAiSseParser from openai_compat
