+++ tests/zai_test.rs
@@ -1,6 +1,6 @@
 #[test]
 fn test_parse_sse_text_delta() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"delta":{"content":"Hello"}}]}"#;
     let event = parser.parse(data).unwrap();
     match event {
@@ -13,7 +13,7 @@ fn test_parse_sse_text_delta() {
 
 #[test]
 fn test_parse_sse_done_event() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = "[DONE]";
     let event = parser.parse(data).unwrap();
     match event {
@@ -24,7 +24,7 @@ fn test_parse_sse_done_event() {
 
 #[test]
 fn test_parse_sse_empty_delta_returns_none() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"delta":{}}]}"#;
     let event = parser.parse(data).unwrap();
     assert!(
@@ -35,7 +35,7 @@ fn test_parse_sse_empty_delta_returns_none() {
 
 #[test]
 fn test_parse_sse_empty_content_returns_none() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"delta":{"content":""}}]}"#;
     let event = parser.parse(data).unwrap();
     assert!(event.is_none(), "empty content string should be ignored");
@@ -43,7 +43,7 @@ fn test_parse_sse_empty_content_returns_none() {
 
 #[test]
 fn test_parse_sse_malformed_json_returns_err() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = "this is not json";
     let result = parser.parse(data);
     assert!(result.is_err(), "malformed JSON should return Err");
@@ -51,7 +51,7 @@ fn test_parse_sse_malformed_json_returns_err() {
 
 #[test]
 fn test_parse_sse_tool_use_start() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"delta":{"tool_calls":[{"index":0,"function":{"name":"bash","arguments":""}}]}}]}"#;
     let event = parser.parse(data).unwrap();
     match event {
@@ -65,7 +65,7 @@ fn test_parse_sse_tool_use_start() {
 
 #[test]
 fn test_parse_sse_tool_use_delta() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data =
         r#"{"choices":[{"delta":{"tool_calls":[{"index":0,"function":{"arguments":"ls"}}]}}]}"#;
     let event = parser.parse(data).unwrap();
@@ -79,7 +79,7 @@ fn test_parse_sse_tool_use_delta() {
 
 #[test]
 fn test_parse_sse_tool_use_done() {
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"finish_reason":"tool_calls"}]}"#;
     let event = parser.parse(data).unwrap();
     match event {
@@ -90,11 +90,9 @@ fn test_parse_sse_tool_use_done() {
 
 #[test]
 fn test_parse_sse_multiple_tool_calls_in_single_chunk() {
-    // Test the critical fix: multiple tool calls in a single chunk are buffered
-    let mut parser = illustrious_manager::backend::zai::ZaiSseParser::new();
+    let mut parser = illustrious_manager::backend::openai_compat::OpenAiSseParser::new();
     let data = r#"{"choices":[{"delta":{"tool_calls":[{"index":0,"function":{"name":"bash","arguments":""}},{"index":1,"function":{"name":"read_file","arguments":""}}]}}]}"#;
 
-    // First call returns first tool call
     let event1 = parser.parse(data).unwrap();
     match event1 {
         Some(illustrious_manager::types::StreamEvent::ToolUseStart { name, .. }) => {
@@ -103,7 +101,6 @@ fn test_parse_sse_multiple_tool_calls_in_single_chunk() {
         other => panic!("Expected ToolUseStart for bash, got {:?}", other),
     }
 
-    // Second call returns second tool call (from buffer)
     let event2 = parser.parse("").unwrap();
     match event2 {
         Some(illustrious_manager::types::StreamEvent::ToolUseStart { name, .. }) => {
@@ -112,7 +109,6 @@ fn test_parse_sse_multiple_tool_calls_in_single_chunk() {
         other => panic!("Expected ToolUseStart for read_file, got {:?}", other),
     }
 
-    // Third call returns None (buffer empty)
     let event3 = parser.parse("").unwrap();
     assert!(event3.is_none(), "Expected None when buffer is empty");
 }
